feat: prompt context
All the prompts right now sort of had the limitation that they were not aware of how specifications or implementations should have been done.

A naive approach would have been to simply make it so that the original specification would be required as a dependency on all the other specifications. This adds a level of coloring to the specification that I don't want people to have to deal with, so instead it should be delegated to the prompts. As people might want to plug and play other specification standards. Currently other specification standards are out of scope for this project, but I don't want to intentionally make it harder for people.