ref(specman-core): change naming
Since the original specification was created with a prompt for making a library, it created a lot of unnecessary parts. This commit aims to remove those parts and functionality so that it doesn't confuse the AI when creating an implementation as the library output should be an implementation detail and not part of the specification.