feat(specman-library): metadata mutation
The intent of this commit is to allow for consumers of the library to simply change the metadata of a field or random artifact without having to make some custom implementation. Ideally, for implementers, this should be a system where anyone using your library should automatically just be having to do this in the first place, as it prevents bad state.