ref(markdown-templates): back-and-forth prompts
When creating the actual prompts, it needs to be made so that the AI responds back to the user that initiated the prompt, stating if there are actions that require user intervention. To do this, I decided to update the prompts so that it includes a very explicit task saying to go through and for each open-ended question presented in bulk to the user where the user can respond in bulk as well.