rev(specman-data-model): dependencies, cycles, scratch pads
This commit contains clarifications for how Scratchpads should be handled when being deleted as well as adding dependencies to the Scratchpads themselves.
While I was at it, I went ahead and just added some more clarifications on what the dependencies and references should be and how their behavior should act just so that there's no confusion.