ref(markdown-templates): add fix, better prompts
One of the things I wanted to improve with the current templates and the prompts is of two different scenarios. The first one is that for the actual specification templates, they had a generic constraints area, which the AI was putting stuff into. I wanted to make it so that these constraints were put under the related entities and concepts, as that should really be how it should be in the first place.
The second is regarding the implementations, and what I wanted to improve there was making it so that the actual implementation documents more closely relate to the specifications that they're implementing. Right now, it had an API surface section, which was kind of vague and never actually used, so I decided to rewrite these prompts so that it was more explicitly worded that API signatures should be placed under sections related to the specification parts that they're implementing.
And then finally, just adding the fix prompts and template, since that didn't exist yet.