ref(markdown-templates): better template designs
The previous designs were very much meant to sort of allow you to copy and paste and then manually fill in info... Because that was my current workflow up to this point. Now, at this point, the prompts are decidedly meant for the MCP server. So, that's the first priority here. The prompts, if they don't work when you're manually pasting it in, is fine with me.