ref(specman-mcp-rust): change back to old workflow
This new workflow was an attempt to try and prevent recency bias. It doesn't seem to have done a good job at it, so as a result I'm just looking to switch back. It's still using sampling for name generation, as that will make it easier for the AI to follow sequential instructions.