rev(specman-core): specman initial draft
This contains the initial draft for structure and relationships.
The initial purpose of this structure was intended for GraphRAG, but deterministic tooling may provide us the means to do TDD, though the means which the TDD is implemented may be loose, as it would ultimately still be up to AI implementation.