ref: templates inside of related directories
Right now, the current system with symlinks sucks because of Windows being awful to use.
I know it's an awful idea to bend backwards for shitty Microsoft products, but even then I still want this to work so more people can contribute and use this tool, so this just cleans up how the markdown templates are stored so that they are not stored using symlinks.